Implemented save/restore states! (*Note slightly bugged because of getParameter implementation*)
Implemented adding/restoring states! Please note that there’s a bug
with the getParameters function that makes the restoration not work.

HUGE NOTE:
Please make sure to change the setParameters function so that it’s
GameViewController-agnostic! We stow away from modularity when we make
things that depend on other files! Make it more reusable; it should be
a general-purpose function!