Shape to Sprite
Updated the code to use SKSpriteNode instead of SKShapeNode objects.
Made it so that one function creates every object we need. Added more
objects for users to play around with. The object no longer turns black
when the user clicks on it though, because sprites adopt the colors of
the image. We should make it so that the selected shape is highlighted.