Joining two objects
Joins two object together if they touch. If it’s paused, you can still
remove the two objects. The joining is kinda ratchet, so it needs to be
modified. If you want to see it work, place one object right next to
another at the start of the simulation and hit the play button
afterwards. They should fall down together. It makes the physics engine
kind of buggy.